Initial/Final commit
Created an event listener that loads the anonymous function when the page is fully loaded. Getting All select elements thats nested in the govLinks. Looping through all the links and every time a link is clicked on the page will load the link on a new page. Changed the file names, Added script link to the js on the html file. Ready for commit.